Move MeetSession features to separate sub-feature
