+++ app/src/main/kotlin/me/nemiron/khinkalyator/features/meets/DI.kt
@@ -9,17 +9,17 @@ import me.nemiron.khinkalyator.features.meets.domain.CreateMeetUseCase
 import me.nemiron.khinkalyator.features.meets.domain.MeetId
 import me.nemiron.khinkalyator.features.meets.domain.MeetsStorage
 import me.nemiron.khinkalyator.features.meets.domain.ObserveMeetsUseCase
-import me.nemiron.khinkalyator.features.meets.domain.ObserveMeetSessionUseCase
+import me.nemiron.khinkalyator.features.meets.session.domain.ObserveMeetSessionUseCase
 import me.nemiron.khinkalyator.features.meets.home_page.ui.MeetsPageComponent
 import me.nemiron.khinkalyator.features.meets.home_page.ui.RealMeetsPageComponent
-import me.nemiron.khinkalyator.features.meets.meet_session_check.ui.MeetSessionCheckComponent
-import me.nemiron.khinkalyator.features.meets.meet_session_check.ui.RealMeetSessionCheckComponent
-import me.nemiron.khinkalyator.features.meets.meet_session_details.ui.MeetSessionDetailsComponent
-import me.nemiron.khinkalyator.features.meets.meet_session_details.ui.RealMeetSessionDetailsComponent
-import me.nemiron.khinkalyator.features.meets.meet_session_overview.ui.MeetSessionOverviewComponent
-import me.nemiron.khinkalyator.features.meets.meet_session_overview.ui.RealMeetSessionOverviewComponent
-import me.nemiron.khinkalyator.features.meets.meet_session_pager.ui.MeetSessionPagerComponent
-import me.nemiron.khinkalyator.features.meets.meet_session_pager.ui.RealMeetSessionPagerComponent
+import me.nemiron.khinkalyator.features.meets.session.check.ui.MeetSessionCheckComponent
+import me.nemiron.khinkalyator.features.meets.session.check.ui.RealMeetSessionCheckComponent
+import me.nemiron.khinkalyator.features.meets.session.details.ui.MeetSessionDetailsComponent
+import me.nemiron.khinkalyator.features.meets.session.details.ui.RealMeetSessionDetailsComponent
+import me.nemiron.khinkalyator.features.meets.session.overview.ui.MeetSessionOverviewComponent
+import me.nemiron.khinkalyator.features.meets.session.overview.ui.RealMeetSessionOverviewComponent
+import me.nemiron.khinkalyator.features.meets.session.pager.ui.MeetSessionPagerComponent
+import me.nemiron.khinkalyator.features.meets.session.pager.ui.RealMeetSessionPagerComponent
 import org.koin.core.component.get
 import org.koin.core.module.dsl.bind
 import org.koin.core.module.dsl.factoryOf
