+++ app/src/main/kotlin/me/nemiron/khinkalyator/features/meets/session/overview/ui/MeetSessionOverviewComponent.kt
@@ -1,9 +1,9 @@
-package me.nemiron.khinkalyator.features.meets.meet_session_overview.ui
+package me.nemiron.khinkalyator.features.meets.session.overview.ui
 
 import com.arkivanov.decompose.router.stack.ChildStack
-import me.nemiron.khinkalyator.features.meets.meet_session_check.ui.MeetSessionCheckComponent
-import me.nemiron.khinkalyator.features.meets.meet_session_details.ui.MeetSessionDetailsComponent
-import me.nemiron.khinkalyator.features.meets.meet_session_pager.ui.MeetSessionPagerComponent
+import me.nemiron.khinkalyator.features.meets.session.check.ui.MeetSessionCheckComponent
+import me.nemiron.khinkalyator.features.meets.session.details.ui.MeetSessionDetailsComponent
+import me.nemiron.khinkalyator.features.meets.session.pager.ui.MeetSessionPagerComponent
 
 interface MeetSessionOverviewComponent {
