+++ app/src/main/kotlin/me/nemiron/khinkalyator/root/ui/RealRootComponent.kt
@@ -19,7 +19,7 @@ import me.nemiron.khinkalyator.features.meets.create.ui.CreateMeetComponent
 import me.nemiron.khinkalyator.features.meets.createCreateMeetComponent
 import me.nemiron.khinkalyator.features.meets.createMeetSessionOverviewComponent
 import me.nemiron.khinkalyator.features.meets.domain.MeetId
-import me.nemiron.khinkalyator.features.meets.meet_session_overview.ui.MeetSessionOverviewComponent
+import me.nemiron.khinkalyator.features.meets.session.overview.ui.MeetSessionOverviewComponent
 import me.nemiron.khinkalyator.features.restaraunts.createRestaurantOverviewComponent
 import me.nemiron.khinkalyator.features.restaraunts.restaurant_overview.ui.RestaurantOverviewComponent
